Config

+++ transaction.go
@@ -5,7 +5,7 @@ import "time"
 type Transaction struct {
 	Symbol        string
 	Owner         string
-	Shares        int
+	Shares        float64
 	PricePerShare float64
 	Type          string // Acquisition or Disposition (A or D)
 	Date          time.Time
